More bug fixes.

+++ ManuvrOS/Transports/BufferPipes/ManuvrTLS/ManuvrTLS.cpp
@@ -75,7 +75,7 @@ extern "C" {
 ManuvrTLS::ManuvrTLS(BufferPipe* _n, int debug_lvl) : BufferPipe() {
   _bp_set_flag(BPIPE_FLAG_IS_BUFFERED, true);
   // mbedTLS will expect this array to be null-terminated. Zero it all...
-  for (uint16_t x = 0; x < MAX_CIPHERSUITE_COUNT; x++) {
+  for (int x = 0; x < MAX_CIPHERSUITE_COUNT; x++) {
     allowed_ciphersuites[x] = 0;
   }
   mbedtls_ssl_config_init(&_conf);
